requirement.txt
Django
numpy
Pillow
questplus
psychopy
pandas
psychometric